Separate App.js code into components
