+++ section9/project1/src/components/Header.js
@@ -1,0 +1,13 @@
+import logo from "../assets/investment-calculator-logo.png";
+import styles from "./Header.module.css";
+
+const Header = () => {
+  return (
+    <header className={styles.header}>
+      <img src={logo} alt="logo" />
+      <h1>Investment Calculator</h1>
+    </header>
+  );
+};
+
+export default Header;
