+++ section9/project1/src/components/Header.module.css
@@ -1,0 +1,15 @@
+.header {
+  text-align: center;
+  margin: 3rem auto;
+}
+
+.header img {
+  width: 5rem;
+  height: 5rem;
+  object-fit: contain;
+  background-color: transparent;
+}
+
+.header h1 {
+  font-size: 1.5rem;
+}
