+++ section9/project1/src/components/InputForm.js
@@ -1,0 +1,40 @@
+import styles from "./InputForm.module.css";
+
+const InputForm = () => {
+  return (
+    <form className={styles.form}>
+      <div className={styles["input-group"]}>
+        <p>
+          <label htmlFor="current-savings">Current Savings ($)</label>
+          <input type="number" id="current-savings" />
+        </p>
+        <p>
+          <label htmlFor="yearly-contribution">Yearly Savings ($)</label>
+          <input type="number" id="yearly-contribution" />
+        </p>
+      </div>
+      <div className={styles["input-group"]}>
+        <p>
+          <label htmlFor="expected-return">
+            Expected Interest (%, per year)
+          </label>
+          <input type="number" id="expected-return" />
+        </p>
+        <p>
+          <label htmlFor="duration">Investment Duration (years)</label>
+          <input type="number" id="duration" />
+        </p>
+      </div>
+      <p className={styles.actions}>
+        <button type="reset" className={styles.buttonAlt}>
+          Reset
+        </button>
+        <button type="submit" className={styles.button}>
+          Calculate
+        </button>
+      </p>
+    </form>
+  );
+};
+
+export default InputForm;
